Refactored and added documention.
Change-list:
Migrated from int to float for movement processing. This allows game objects to move at speeds that are not divisible by the framerate.
Added Rectf, a struct that holds float x,y,w,h values.
EVENT_SCORED is to be depreciated, as the current mechanism for scoring does not rely on event handling. Using it invokes undefined behavior because it is not assigned a proper value.
Added various documentation.